# docs/swagger.yaml
basePath: /
definitions:
  domain.Comment:
    properties:
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      photo_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  domain.CommentAdd:
    properties:
      message:
        type: string
      photo_id:
        type: integer
    required:
    - message
    - photo_id
    type: object
  domain.CommentAddResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      photo_id:
        type: integer
      updated_at:
        type: string
    type: object
  domain.CommentDeleteResponse:
    properties:
      message:
        type: string
    type: object
  domain.CommentUpdateData:
    properties:
      message:
        type: string
    required:
    - message
    type: object
  domain.CommentUpdateDataResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      photo_id:
        type: integer
      updated_at:
        type: string
    type: object
  domain.CommentWithUserAndPhotoData:
    properties:
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      photo:
        $ref: '#/definitions/domain.Photo'
      photo_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/domain.User'
      user_id:
        type: integer
    type: object
  domain.Photo:
    properties:
      caption:
        type: string
      comment:
        items:
          $ref: '#/definitions/domain.Comment'
        type: array
      created_at:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  domain.PhotoAdd:
    properties:
      caption:
        type: string
      photo_url:
        type: string
      title:
        type: string
    required:
    - photo_url
    - title
    type: object
  domain.PhotoAddResponse:
    properties:
      caption:
        type: string
      created_at:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  domain.PhotoDeleteResponse:
    properties:
      message:
        type: string
    type: object
  domain.PhotoUpdateData:
    properties:
      caption:
        type: string
      photo_url:
        type: string
      title:
        type: string
    required:
    - caption
    - photo_url
    - title
    type: object
  domain.PhotoUpdateDataResponse:
    properties:
      caption:
        type: string
      created_at:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  domain.PhotoWithUserData:
    properties:
      caption:
        type: string
      comment:
        items:
          $ref: '#/definitions/domain.Comment'
        type: array
      created_at:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/domain.User'
      user_id:
        type: integer
    type: object
  domain.SocialMedia:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      social_media_url:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  domain.User:
    properties:
      age:
        type: integer
      comments:
        items:
          $ref: '#/definitions/domain.Comment'
        type: array
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      photo:
        $ref: '#/definitions/domain.Photo'
      social_media:
        $ref: '#/definitions/domain.SocialMedia'
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  middleware.Error:
    properties:
      message:
        type: string
    type: object
host: localhost:8001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample gin server for managing user's photos and social medias
    with Authentication & Authorization.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Photos & Social Meida Management API
  version: "1.0"
paths:
  /comments:
    get:
      consumes:
      - application/json
      parameters:
      - description: Access Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.CommentWithUserAndPhotoData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.Error'
      summary: Get all comments
      tags:
      - Photo's Comments
    post:
      consumes:
      - application/json
      parameters:
      - description: Access Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Comment Data
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/domain.CommentAdd'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.CommentAddResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.Error'
      summary: Add comment to existing photo
      tags:
      - Photo's Comments
  /comments/{commentId}:
    delete:
      parameters:
      - description: Access Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.CommentDeleteResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.Error'
      summary: Delete existing comment
      tags:
      - Photo's Comments
    put:
      consumes:
      - application/json
      parameters:
      - description: Access Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: integer
      - description: Comment Data
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/domain.CommentUpdateData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.CommentUpdateDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.Error'
      summary: Update existing comment
      tags:
      - Photo's Comments
  /photos:
    get:
      parameters:
      - description: Access Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.PhotoWithUserData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.Error'
      summary: Get all photos
      tags:
      - Photos
    post:
      consumes:
      - application/json
      parameters:
      - description: Access Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Photo Data
        in: body
        name: photo
        required: true
        schema:
          $ref: '#/definitions/domain.PhotoAdd'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.PhotoAddResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.Error'
      summary: Add a new user photo
      tags:
      - Photos
  /photos/{photoId}:
    delete:
      parameters:
      - description: Access Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Photo ID
        in: path
        name: photoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.PhotoDeleteResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.Error'
      summary: Delete existing photo
      tags:
      - Photos
    put:
      consumes:
      - application/json
      parameters:
      - description: Access Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Photo ID
        in: path
        name: photoId
        required: true
        type: integer
      - description: Photo Data
        in: body
        name: photo
        required: true
        schema:
          $ref: '#/definitions/domain.PhotoUpdateData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.PhotoUpdateDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.Error'
      summary: Update existing photo
      tags:
      - Photos
...
swagger: "2.0"